+++ src/main/java/com/github/yunabraska/githubworkflow/services/HighlightAnnotator.java
@@ -11,8 +11,10 @@ import com.intellij.lang.annotation.Annotator;
 import com.intellij.lang.annotation.HighlightSeverity;
 import com.intellij.openapi.editor.DefaultLanguageHighlighterColors;
 import com.intellij.openapi.keymap.KeymapUtil;
+import com.intellij.openapi.util.Key;
 import com.intellij.openapi.util.TextRange;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.impl.source.tree.LeafPsiElement;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.yaml.psi.YAMLKeyValue;
 import org.jetbrains.yaml.psi.YAMLSequenceItem;
@@ -25,8 +27,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -40,12 +42,12 @@ import static java.util.Optional.ofNullable;
 
 public class HighlightAnnotator implements Annotator {
 
-    public static final Pattern CARET_BRACKET_ITEM_PATTERN = Pattern.compile("[\\s|\\t^{]\\b(\\w++(?:\\.\\w++)++)[\\s|\\t$}]");
-
     @Override
     public void annotate(@NotNull final PsiElement psiElement, @NotNull final AnnotationHolder holder) {
         if (psiElement.isValid()) {
             processPsiElement(holder, psiElement);
+            variableElementHandler(holder, psiElement);
+            highlightRunOutputs(holder, psiElement);
         }
     }
 
@@ -55,42 +57,30 @@ public class HighlightAnnotator implements Annotator {
                 case FIELD_USES -> usesHandler(holder, element);
                 case FIELD_WITH -> withHandler(holder, element);
                 case FIELD_NEEDS -> needsHandler(holder, element);
-                case FIELD_RUN -> runHandler(holder, element);
                 case FIELD_OUTPUTS -> outputsHandler(holder, element);
                 default -> {
                     // No Action
                 }
             }
         });
-
-        // action.yaml
-//        ofNullable(psiElement)
-//                .filter(LeafPsiElement.class::isInstance)
-//                .map(LeafPsiElement.class::cast)
-//                .map(PsiElement::getParent)
-//                .filter(YAMLKeyValue.class::isInstance)
-//                .map(YAMLKeyValue.class::cast)
-//                .filter(with -> FIELD_WITH.equals(with.getKeyText()))
-//                .filter(with -> psiElement.getText().equals(FIELD_WITH))
-//                .ifPresent(with -> withHandler(holder, psiElement));
-
-        // GENERAL
-        variableHandler(holder, psiElement);
     }
 
-    private static void runHandler(final AnnotationHolder holder, final YAMLKeyValue element) {
+    private static void highlightRunOutputs(final AnnotationHolder holder, final PsiElement psiElement) {
         // SHOW Output Env && Output Variable declaration
-        Stream.of(
-                parseEnvVariables(element),
-                parseOutputVariables(element)
-        ).flatMap(Collection::stream).collect(Collectors.groupingBy(SimpleElement::startIndexOffset)).forEach((integer, elements) -> ofNullable(getFirstChild(elements)).ifPresent(lineElement -> holder
-                .newSilentAnnotation(INFORMATION)
-                .range(lineElement.range())
-                .gutterIconRenderer(new IconRenderer(null, element, ICON_TEXT_VARIABLE))
-                .create()
-        ));
+        Optional.of(psiElement)
+                .filter(LeafPsiElement.class::isInstance)
+                .map(LeafPsiElement.class::cast)
+                .filter(element -> PsiElementHelper.getParent(element, FIELD_RUN).isPresent())
+                .ifPresent(element -> Stream.of(
+                        parseEnvVariables(element),
+                        parseOutputVariables(element)
+                ).flatMap(Collection::stream).collect(Collectors.groupingBy(SimpleElement::startIndexOffset)).forEach((integer, elements) -> ofNullable(getFirstChild(elements)).ifPresent(lineElement -> holder
+                        .newSilentAnnotation(INFORMATION)
+                        .range(lineElement.range())
+                        .gutterIconRenderer(new IconRenderer(null, element, ICON_TEXT_VARIABLE))
+                        .create()
+                )));
     }
-
 
     @SuppressWarnings("DataFlowIssue")
     private static void usesHandler(final AnnotationHolder holder, final YAMLKeyValue element) {
@@ -190,73 +180,157 @@ public class HighlightAnnotator implements Annotator {
     }
 
 
-    private static void variableHandler(final AnnotationHolder holder, final PsiElement psiElement) {
-        getParent(psiElement, FIELD_RUN)
-                .or(() -> getParent(psiElement, FIELD_ID))
-                .or(() -> getParent(psiElement, "name"))
-                .or(() -> getParent(psiElement, "value").isPresent() ? getParent(psiElement, FIELD_OUTPUTS) : Optional.empty())
-                .or(() -> getParent(psiElement, FIELD_WITH))
-                .or(() -> getParent(psiElement, FIELD_ENVS))
-                .or(() -> getParent(psiElement, FIELD_OUTPUTS))
-                .ifPresent(parent -> {
-                    final Matcher matcher = CARET_BRACKET_ITEM_PATTERN.matcher(psiElement.getText());
-                    while (matcher.find()) {
-                        final String[] parts = Arrays.stream(matcher.group().split("\\."))
-                                .map(s -> s.replace("{", ""))
-                                .map(s -> s.replace("}", ""))
-                                .map(String::trim)
-                                .toArray(String[]::new);
-                        final String scope = parts[0];
-                        switch (scope) {
-                            case FIELD_INPUTS ->
-                                    ifEnoughItems(holder, psiElement, parts, 2, 2, inputId -> isDefinedItem0(psiElement, holder, matcher, inputId, listInputs(psiElement).stream().map(SimpleElement::key).toList()));
-                            case FIELD_SECRETS -> ifEnoughItems(holder, psiElement, parts, 2, 2, secretId -> {
-                                final List<String> secrets = listSecrets(psiElement).stream().map(SimpleElement::key).toList();
-                                if (!secrets.contains(secretId)) {
-                                    final TextRange textRange = simpleTextRange(psiElement, matcher, secretId);
-                                    createAnnotation(psiElement, textRange, holder, secrets.stream().map(secret -> new SyntaxAnnotation(
-                                            "Replace [" + secretId + "] with [" + secret + "] - if it is not provided at runtime",
-                                            null,
-                                            HighlightSeverity.WEAK_WARNING,
-                                            ProblemHighlightType.WEAK_WARNING,
-                                            replaceAction(textRange, secret),
-                                            true
-                                    )).toList());
-                                }
-                            });
-                            case FIELD_ENVS ->
-                                    ifEnoughItems(holder, psiElement, parts, 2, -1, envId -> isDefinedItem0(psiElement, holder, matcher, envId, listEnvs(psiElement).stream().map(SimpleElement::key).toList()));
-                            case FIELD_GITHUB ->
-                                    ifEnoughItems(holder, psiElement, parts, 2, -1, envId -> isDefinedItem0(psiElement, holder, matcher, envId, new ArrayList<>(DEFAULT_VALUE_MAP.get(FIELD_GITHUB).get().keySet())));
-                            case FIELD_RUNNER ->
-                                    ifEnoughItems(holder, psiElement, parts, 2, 2, runnerId -> isDefinedItem0(psiElement, holder, matcher, runnerId, new ArrayList<>(DEFAULT_VALUE_MAP.get(FIELD_RUNNER).get().keySet())));
-                            case FIELD_STEPS -> ifEnoughItems(holder, psiElement, parts, 4, 4, stepId -> {
-                                final List<YAMLSequenceItem> steps = listSteps(psiElement);
-                                if (isDefinedItem0(psiElement, holder, matcher, stepId, steps.stream().map(step -> getText(step, FIELD_ID).orElse(null)).filter(Objects::nonNull).toList()) && isField2Valid(psiElement, holder, matcher, parts[2])) {
-                                    final List<String> outputs = listStepOutputs(steps.stream().filter(step -> getText(step, FIELD_ID).filter(id -> id.equals(stepId)).isPresent()).findFirst().orElse(null)).stream().map(SimpleElement::key).toList();
-                                    isValidItem3(psiElement, holder, matcher, parts[3], outputs);
+    private static void variableElementHandler(final AnnotationHolder holder, final PsiElement psiElement) {
+        final Optional<YAMLKeyValue> parentIf = getParent(psiElement, FIELD_IF);
+        Optional.of(psiElement)
+                .filter(LeafPsiElement.class::isInstance)
+                .map(LeafPsiElement.class::cast)
+                .filter(isElementWithVariables(parentIf.orElse(null)))
+                .ifPresent(element -> toSimpleElements(element).forEach(simpleElement -> {
+                            final SimpleElement[] parts = splitToElements(simpleElement);
+                            switch (parts.length > 0 ? parts[0].text() : "N/A") {
+                                case FIELD_INPUTS ->
+                                        ifEnoughItems(holder, element, parts, 2, 2, inputId -> isDefinedItem0(element, holder, inputId, listInputs(element).stream().map(SimpleElement::key).toList()));
+                                case FIELD_SECRETS -> ifEnoughItems(holder, element, parts, 2, 2, secretId -> {
+                                    // SECRETS ARE NOT ALLOWED IN IF STATEMENT
+                                    if (parentIf.isPresent()) {
+                                        final TextRange range = psiElement.getTextRange();
+                                        final TextRange textRange = new TextRange(range.getStartOffset() + parts[0].startIndexOffset(), range.getStartOffset() + parts[parts.length - 1].endIndexOffset());
+                                        new SyntaxAnnotation(
+                                                "Remove [" + simpleElement.text() + "] - Secrets are not valid in `if` statements",
+                                                null,
+                                                deleteElementAction(textRange)
+                                        ).createAnnotation(psiElement, textRange, holder);
+                                    }
+                                    final List<String> secrets = listSecrets(element).stream().map(SimpleElement::key).toList();
+                                    if (!secrets.contains(secretId.text())) {
+                                        final TextRange textRange = simpleTextRange(element, secretId);
+                                        createAnnotation(element, textRange, holder, secrets.stream().map(secret -> new SyntaxAnnotation(
+                                                "Replace [" + secretId.text() + "] with [" + secret + "] - if it is not provided at runtime",
+                                                null,
+                                                HighlightSeverity.WEAK_WARNING,
+                                                ProblemHighlightType.WEAK_WARNING,
+                                                replaceAction(textRange, secret),
+                                                true
+                                        )).toList());
+                                    }
+                                });
+                                case FIELD_ENVS ->
+                                        ifEnoughItems(holder, element, parts, 2, -1, envId -> isDefinedItem0(element, holder, envId, listEnvs(element).stream().map(SimpleElement::key).toList()));
+                                case FIELD_GITHUB ->
+                                        ifEnoughItems(holder, element, parts, 2, -1, envId -> isDefinedItem0(element, holder, envId, new ArrayList<>(DEFAULT_VALUE_MAP.get(FIELD_GITHUB).get().keySet())));
+                                case FIELD_RUNNER ->
+                                        ifEnoughItems(holder, element, parts, 2, 2, runnerId -> isDefinedItem0(element, holder, runnerId, new ArrayList<>(DEFAULT_VALUE_MAP.get(FIELD_RUNNER).get().keySet())));
+                                case FIELD_STEPS -> ifEnoughItems(holder, element, parts, 4, 4, stepId -> {
+                                    final List<YAMLSequenceItem> steps = listSteps(element);
+                                    if (isDefinedItem0(element, holder, stepId, steps.stream().map(step -> getText(step, FIELD_ID).orElse(null)).filter(Objects::nonNull).toList()) && isField2Valid(element, holder, parts[2])) {
+                                        final List<String> outputs = listStepOutputs(steps.stream().filter(step -> getText(step, FIELD_ID).filter(id -> id.equals(stepId.text())).isPresent()).findFirst().orElse(null)).stream().map(SimpleElement::key).toList();
+                                        isValidItem3(element, holder, parts[3], outputs);
 
+                                    }
+                                });
+                                case FIELD_JOBS -> ifEnoughItems(holder, element, parts, 4, 4, jobId -> {
+                                    final List<YAMLKeyValue> jobs = listJobs(element);
+                                    if (isDefinedItem0(element, holder, jobId, jobs.stream().map(YAMLKeyValue::getKeyText).toList()) && isField2Valid(element, holder, parts[2])) {
+                                        final List<String> outputs = listJobOutputs(jobs.stream().filter(job -> job.getKeyText().equals(jobId.text())).findFirst().orElse(null)).stream().map(SimpleElement::key).toList();
+                                        isValidItem3(element, holder, parts[3], outputs);
+                                    }
+                                });
+                                case FIELD_NEEDS -> ifEnoughItems(holder, element, parts, 4, 4, jobId -> {
+                                    final List<String> jobIds = listJobNeeds(element);
+                                    if (isDefinedItem0(element, holder, jobId, jobIds) && isField2Valid(element, holder, parts[2])) {
+                                        final List<String> outputs = listJobOutputs(listAllJobs(element).stream().filter(job -> job.getKeyText().equals(jobId.text())).findFirst().orElse(null)).stream().map(SimpleElement::key).toList();
+                                        isValidItem3(element, holder, parts[3], outputs);
+                                    }
+                                });
+                                default -> {
+                                    // ignored
                                 }
-                            });
-                            case FIELD_JOBS -> ifEnoughItems(holder, psiElement, parts, 4, 4, jobId -> {
-                                final List<YAMLKeyValue> jobs = listJobs(psiElement);
-                                if (isDefinedItem0(psiElement, holder, matcher, jobId, jobs.stream().map(YAMLKeyValue::getKeyText).toList()) && isField2Valid(psiElement, holder, matcher, parts[2])) {
-                                    final List<String> outputs = listJobOutputs(jobs.stream().filter(job -> jobId.equals(job.getKeyText())).findFirst().orElse(null)).stream().map(SimpleElement::key).toList();
-                                    isValidItem3(psiElement, holder, matcher, parts[3], outputs);
-                                }
-                            });
-                            case FIELD_NEEDS -> ifEnoughItems(holder, psiElement, parts, 4, 4, jobId -> {
-                                final List<String> jobIds = listJobNeeds(psiElement);
-                                if (isDefinedItem0(psiElement, holder, matcher, jobId, jobIds) && isField2Valid(psiElement, holder, matcher, parts[2])) {
-                                    final List<String> outputs = listJobOutputs(listAllJobs(psiElement).stream().filter(job -> jobId.equals(job.getKeyText())).findFirst().orElse(null)).stream().map(SimpleElement::key).toList();
-                                    isValidItem3(psiElement, holder, matcher, parts[3], outputs);
-                                }
-                            });
-                            default -> {
-                                // ignored
                             }
-                        }
-                    }
-                });
+                        })
+                );
+    }
+
+    @NotNull
+    public static Predicate<PsiElement> isElementWithVariables(final YAMLKeyValue parentIf) {
+        return element -> ofNullable(parentIf)
+                .or(() -> getParent(element, FIELD_RUN))
+                .or(() -> getParent(element, FIELD_ID))
+                .or(() -> getParent(element, "name"))
+                .or(() -> getParent(element, "value").isPresent() ? getParent(element, FIELD_OUTPUTS) : Optional.empty())
+                .or(() -> getParent(element, FIELD_WITH))
+                .or(() -> getParent(element, FIELD_ENVS))
+                .or(() -> getParent(element, FIELD_OUTPUTS))
+                .isPresent();
+    }
+
+    public static Key<SimpleElement[]> VARIABLE_ELEMENTS = new Key<>("com.github.yunabraska.githubworkflow.VariableElements");
+
+    @NotNull
+    public static List<SimpleElement> toSimpleElements(final PsiElement element) {
+        return Arrays.stream(element.getText().split("\\R"))
+                .map(String::trim).filter(PsiElementHelper::hasText)
+                // EXCLUDE COMMENT LINES
+                .filter(s -> !s.startsWith("#"))
+                .flatMap(s -> findDottedExpressions(s).stream())
+                .toList();
+    }
+
+    @NotNull
+    public static SimpleElement[] splitToElements(final SimpleElement simpleElement) {
+        final AtomicInteger start = new AtomicInteger(simpleElement.range().getStartOffset());
+        return Arrays.stream(simpleElement.text().split("\\."))
+                .map(s -> s.replace("{", ""))
+                .map(s -> s.replace("}", ""))
+                .map(String::trim)
+                .filter(PsiElementHelper::hasText)
+                .map(part -> {
+                    final int length = part.length();
+                    final TextRange range = new TextRange(start.get(), start.get() + length);
+                    start.addAndGet(length + 1);
+                    return new SimpleElement(part, range);
+                })
+                .toArray(SimpleElement[]::new);
+    }
+
+    public static List<SimpleElement> findDottedExpressions(final String text) {
+        final List<SimpleElement> elements = new ArrayList<>();
+        final StringBuilder currentElement = new StringBuilder();
+        int elementStart = -1;
+        char previousChar = ' ';
+
+        for (int i = 0; i < text.length(); i++) {
+            final char ch = text.charAt(i);
+            final boolean letterOrDigit = Character.isLetterOrDigit(ch);
+
+
+            if (elementStart == -1 && letterOrDigit && (Character.isWhitespace(previousChar) || previousChar != '{')) {
+                // START
+                elementStart = i;
+                currentElement.setLength(0);
+                currentElement.append(ch);
+            } else if (elementStart != -1 && (letterOrDigit || ch == '_' || ch == '-' || ch == '.')) {
+                // MIDDLE
+                currentElement.append(ch);
+                // LAST ITEM
+                if (i + 1 == text.length()) {
+                    elementStart = validateAndAddElement(currentElement, elements, elementStart, i);
+                }
+            } else if (elementStart != -1 && (i + 1 == text.length() || Character.isWhitespace(text.charAt(i)) || text.charAt(i + 1) == '}')) {
+                // END
+                elementStart = validateAndAddElement(currentElement, elements, elementStart, i);
+            }
+            previousChar = ch;
+        }
+        return elements;
+    }
+
+    private static int validateAndAddElement(final StringBuilder currentElement, final List<SimpleElement> elements, int elementStart, final int i) {
+        if (!currentElement.isEmpty() && currentElement.length() > 1 && currentElement.toString().contains(".")) {
+            elements.add(new SimpleElement(currentElement.toString(), new TextRange(elementStart, i)));
+        }
+        elementStart = -1;
+        currentElement.setLength(0);
+        return elementStart;
     }
 }