+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ## [Unreleased]
 
+## [3.0.2] - 2023-10-08
+
+### Refactoring
+
+- Feat: Added autocompletion for `$GITHUB_OUTPUT` and `$GITHUB_ENV`
+- Fix: [#31 StringIndexOutOfBoundsException (GitHubAction.java:87)](https://github.com/YunaBraska/github-workflow-plugin/issues/31)
+- Fix: Secrets are unsupported on if statements 
+- Fix: highlighting issues when workflow is very long 
+- Fix: Code Completion shows only available items [inputs, secrets, jobs, steps, needs]
+- Fix: Exchanged too complex Regex variable matcher with a dedicated function
+
 ## [3.0.1] - 2023-10-01
 
 ### Refactoring